vision/requirements.txt
mediapipe==0.10.14
numpy==1.26.4
opencv-python==4.10.0.84
python-dotenv==1.0.1
simpleaudio==1.0.4
